+++ MicroprocessorLib/ReadProgramBank.cpp
@@ -19,8 +19,7 @@ void CReadProgramBank::Init(NotMemoryLock bNotMemoryLock)
 //////////////////////////////////////////////////////////////////////////
 void CReadProgramBank::Execute(CW65C816* pcCPU)
 {
-	//Is this the right Data?
-	pcCPU->GetState()->SetNewProgramCounterBank((uint8)pcCPU->GetState()->GetData());
+	pcCPU->GetState()->SetNewProgramCounterBank(pcCPU->GetState()->GetData());
 }
